backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
pymongo==4.6.0
pandas==2.3.2
numpy==2.3.2
scikit-learn==1.6.1
python-dotenv==1.0.0
bcrypt==4.0.1
Werkzeug==3.0.1
blinker==1.7.0
joblib==1.5.1
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
PyJWT==2.8.0
scipy==1.16.1
threadpoolctl==3.6.0
